Refactor Observation Scenario tests to share csv code
