+++ src/test/resources/tracking/observation/SurvivalRateT0DensityChange/AfterUpdate/PlotRates.csv
@@ -1,0 +1,5 @@
+Plot Number,Plot Survival Rate,Species A Survival Rate,Species B Survival Rate
+111,73%,45%,100%
+112,99%,100%,98%
+211,100%,110%,92%
+311,93%,87%,98%
