+++ src/test/resources/tracking/observation/SurvivalRateT0DensityChange/AfterUpdate/SiteRates.csv
@@ -1,0 +1,2 @@
+Site Name,Plot Survival Rate,Species A Survival Rate,Species B Survival Rate
+Demo Site,94%,91%,96%
