+++ src/test/resources/tracking/observation/SurvivalRateT0DensityChange/AfterUpdate/ZoneRates.csv
@@ -1,0 +1,3 @@
+Zone Name,Plot Survival Rate,Species A Survival Rate,Species B Survival Rate
+Zone1,95%,94%,95%
+Zone2,93%,87%,98%
