+++ src/test/resources/tracking/observation/SurvivalRateT0DensityChange/Observation-1.csv
@@ -1,0 +1,6 @@
+Plot,Species A,,,Species B,,Unknown,,Other,
+,Existing,Live,Dead,Existing,Live,Dead,Live,Dead,Live,Dead
+111,0,9,0,0,20,0,0,0,0,0
+112,0,30,0,0,39,0,0,0,0,0
+211,0,55,0,0,55,0,0,0,0,0
+311,0,61,0,0,78,0,0,0,0,0
