+++ src/test/resources/tracking/observation/SurvivalRateT0DensityChange/PlotRates.csv
@@ -1,0 +1,5 @@
+monitoring plot id,Plot Survival Rate,Species A Survival Rate,Species B Survival Rate
+111,290%,90%,100%
+112,345%,100%,98%
+211,367%,110%,92%
+311,348%,87%,98%
