+++ src/test/resources/tracking/observation/SurvivalRateT0DensityChange/Plots.csv
@@ -1,0 +1,5 @@
+Subzone,Name,Type,Area (ha)
+Subzone1,111,Permanent,0.0625
+Subzone1,112,Permanent,0.0625
+Subzone2,211,Permanent,0.0625
+Subzone3,311,Permanent,0.0625
