+++ src/test/resources/tracking/observation/SurvivalRateT0DensityChange/SiteRates.csv
@@ -1,0 +1,2 @@
+Site Id,Plot Survival Rate,Species A Survival Rate,Species B Survival Rate
+Demo Site,96%,97%,96%
