+++ src/test/resources/tracking/observation/SurvivalRateT0DensityChange/SubzoneRates.csv
@@ -1,0 +1,4 @@
+Subzone Id,Plot Survival Rate,Species A Survival Rate,Species B Survival Rate
+Subzone1,98%,98%,98%
+Subzone2,100%,110%,92%
+Subzone3,93%,87%,98%
