+++ src/test/resources/tracking/observation/SurvivalRateT0DensityChange/Subzones.csv
@@ -1,0 +1,5 @@
+
+Zone,Name,Area
+Zone1,Subzone1,50
+Zone1,Subzone2,50
+Zone2,Subzone3,200
