+++ src/test/resources/tracking/observation/SurvivalRateT0DensityChange/T0Densities.csv
@@ -1,0 +1,5 @@
+Plot,Species 0 Density,Species 1 Density
+111,10,20
+112,30,40
+211,50,60
+311,70,80
