+++ src/test/resources/tracking/observation/SurvivalRateT0DensityChange/ZoneRates.csv
@@ -1,0 +1,3 @@
+Zone Id,Plot Survival Rate,Species A Survival Rate,Species B Survival Rate
+Zone1,99%,104%,95%
+Zone2,93%,87%,98%
