+++ src/test/resources/tracking/observation/SurvivalRateT0DensityChange/Zones.csv
@@ -1,0 +1,4 @@
+
+Site,Name,Area (ha)
+Demo Site,Zone1,100
+Demo Site,Zone2,200
